added structure for RetornoCnab400
