+++ src/Contracts/Cnab/Retorno/Cnab400/Header.php
@@ -20,117 +20,71 @@
  *   IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-namespace Eduardokum\LaravelBoleto\Contracts\Cnab\Retorno;
+namespace Eduardokum\LaravelBoleto\Contracts\Cnab\Retorno\Cnab400;
 
-interface Detalhe
+interface Header
 {
 
-    const OCORRENCIA_LIQUIDADA = 1;
-    const OCORRENCIA_BAIXADA = 2;
-    const OCORRENCIA_ENTRADA = 3;
-    const OCORRENCIA_ALTERACAO = 4;
-    const OCORRENCIA_PROTESTADA = 5;
-    const OCORRENCIA_OUTROS = 6;
-    const OCORRENCIA_ERRO = 9;
+    /**
+     * @return mixed
+     */
+    public function getOperacaoCodigo();
 
     /**
      * @return mixed
      */
-    public function getNossoNumero();
+    public function getOperacao();
 
     /**
      * @return mixed
      */
-    public function getNumeroDocumento();
+    public function getServicoCodigo();
 
     /**
      * @return mixed
      */
-    public function getOcorrencia();
+    public function getServico();
 
     /**
      * @return mixed
      */
-    public function getOcorrenciaDescricao();
+    public function getAgencia();
 
     /**
      * @return mixed
      */
-    public function getOcorrenciaTipo();
+    public function getAgenciaDigito();
+
+    /**
+     * @return mixed
+     */
+    public function getConta();
+
+    /**
+     * @return mixed
+     */
+    public function getContaDigito();
 
     /**
      * @param string $format
      *
-     * @return mixed
+     * @return \Carbon\Carbon
      */
-    public function getDataOcorrencia($format = 'd/m/Y');
-
-    /**
-     * @param string $format
-     *
-     * @return mixed
-     */
-    public function getDataVencimento($format = 'd/m/Y');
-
-    /**
-     * @param string $format
-     *
-     * @return mixed
-     */
-    public function getDataCredito($format = 'd/m/Y');
+    public function getData($format = 'd/m/Y');
 
     /**
      * @return mixed
      */
-    public function getValor();
+    public function getConvenio();
 
     /**
      * @return mixed
      */
-    public function getValorTarifa();
-
-    /**
-     * @return mixed
-     */
-    public function getValorIOF();
-
-    /**
-     * @return mixed
-     */
-    public function getValorAbatimento();
-
-    /**
-     * @return mixed
-     */
-    public function getValorDesconto();
-
-    /**
-     * @return mixed
-     */
-    public function getValorRecebido();
-
-    /**
-     * @return mixed
-     */
-    public function getValorMora();
-
-    /**
-     * @return mixed
-     */
-    public function getValorMulta();
-
-    /**
-     * @return string
-     */
-    public function getError();
-
-    /**
-     * @return boolean
-     */
-    public function hasError();
+    public function getCodigoCliente();
 
     /**
      * @return array
      */
     public function toArray();
+
 }
