+++ tests/RetornoCnab240Test.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Cnab400\Tests;
+
+use Eduardokum\LaravelBoleto\Cnab\Retorno;
+
+class RetornoCnab240Test extends \PHPUnit_Framework_TestCase
+{
+    public function testRetornoBradescoCnab240()
+    {
+
+
+    }
+}
